[#182] Refactor: usePasswordManagerBox 커스텀 훅
